Fixing cyclic dependency.
